+++ src/MachinaGrasshopper/Utils/MachinaBridgeSocket.cs
@@ -15,7 +15,11 @@ namespace MachinaGrasshopper.Utils
     {
         public WebSocket socket;
         public string name;
-        public List<string> receivedMessage = new List<string>();
+        public List<string> receivedMessages = new List<string>();
+        public int logged = 0;
+
+        public int minCount = 6;
+        public int maxCount = 24;
 
         public MachinaBridgeSocket(string socketName) {
             this.name = socketName;
@@ -23,13 +27,31 @@ namespace MachinaGrasshopper.Utils
 
         public void Log(string msg)
         {
-            receivedMessage.Add(msg);
+            receivedMessages.Add(msg);
+            logged++;
+
+            //ResizeBuffer();
         }
 
         public void Flush()
         {
-            receivedMessage.Clear();
+            receivedMessages.Clear();
+            logged = 0;
         }
+
+        private bool ResizeBuffer()
+        {
+            if (receivedMessages.Count > maxCount)
+            {
+                receivedMessages = receivedMessages.GetRange(receivedMessages.Count - minCount - 1, minCount);
+                logged = receivedMessages.Count;
+                return true;
+            }
+            return false;
+        }
+
+
+
 
     }
 }